Making Comment a part of Track, The Group discussed and instead of having multiple comments per track ( which used the database as Comment Objects) we will have each track have only one comment witch will be ammended as needed.
The Upsert Commands, as aswesome as the concept was, failed because it wouldn't return the primary id of anything new or updated.  Went back to @Insert but i added the onConflict Replace directive to it which makes it function just like upsert, except that it will return the primary key.  We need this because when you create a track then add points you need the trackid to add points, and without having to make calls to db to get all the items and compare things to expected item